更换循迹和解决噪点问题

+++ step_motor/step_motor.h
@@ -5,7 +5,7 @@
 
 void step_motor_Init(void);
 void step_motor_start(uint8_t steeper_id);
-void step_set_speed(uint16_t speed, uint8_t stepper_id);
-void step_set_angle(uint8_t angle, uint8_t stepper_id);
+void step_set_speed(float speed, uint8_t stepper_id);
+void step_set_angle(float angle, uint8_t stepper_id);
 
 #endif
